Capture browser logs as an independent formatter
This way it's available even if classic formatters are not available,
like when using turbo_tests.
