+++ spec/support/capybara.rb
@@ -81,67 +81,3 @@ end
 Rails.application.config do
   config.middleware.use RackSessionAccess::Middleware
 end
-
-# Capture browser logs on failed examples and output them in Progress and
-# Documentation formatters.
-module Capybara::CaptureBrowserLogs
-  RSPEC_TEXT_FORMATTERS = [
-    "RSpec::Core::Formatters::ProgressFormatter",
-    "RSpec::Core::Formatters::DocumentationFormatter"
-  ].freeze
-
-  class << self
-    def after_failed_example(example)
-      return unless example.example_group.include?(Capybara::DSL)
-      return unless failed?(example)
-      return if Capybara.page.current_url.blank?
-      return unless Capybara.page.driver.browser.respond_to?(:manage)
-
-      logs = Capybara.page.driver.browser.manage.instance_variable_get(:@bridge).log("browser")
-      example.metadata[:browser_logs] = logs
-    rescue StandardError => e
-      warn "Unable to get browser logs: #{e}"
-    end
-
-    def change_text_formatter_to_output_captured_browser_logs
-      RSpec.configuration.formatters.each do |formatter|
-        next unless RSPEC_TEXT_FORMATTERS.include?(formatter.class.to_s)
-        next if formatter.singleton_class.included_modules.include?(TextReporter)
-
-        formatter.singleton_class.prepend(TextReporter)
-      end
-    end
-
-    def failed?(example)
-      # reusing private method from capybara-screenshot gem
-      Capybara::Screenshot::RSpec.send(:failed?, example)
-    end
-  end
-
-  module TextReporter
-    def example_failed(notification)
-      super
-      output_browser_logs(notification.example)
-    end
-
-    private
-
-    def output_browser_logs(example)
-      return unless example.metadata[:browser_logs]
-
-      logs = example.metadata[:browser_logs]
-      output.puts("  Browser logs:\n    #{logs.join("\n    ")}")
-    end
-  end
-end
-
-# Output browser logs after failed feature test
-RSpec.configure do |config|
-  config.after(type: :feature) do |example|
-    Capybara::CaptureBrowserLogs.after_failed_example(example)
-  end
-
-  config.before(:suite) do
-    Capybara::CaptureBrowserLogs.change_text_formatter_to_output_captured_browser_logs
-  end
-end
